(feat) Events for LUA scripting
So far so good, they respond properly but it looks like the events are triggered twice (in the case of `onPlayerMessage` the message arg is lost too). I also need to figure out why the IDs aren't incremental but seem random, as well as item instances being numbers instead of text.